+++ esp_video/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Add RISC-V software swap byte function for DVP video device
 - Add the customized video class to call esp_cam_sensor ioctl commands directly
 - Add the second SPI video device
 - Improve AEC controller time precision